% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnadot-beta.R
\name{.dnadot_betabin_internal}
\alias{.dnadot_betabin_internal}
\title{Internal handler for the Beta-binomial method with a fixed, global Fis.}
\usage{
.dnadot_betabin_internal(
  genind_object,
  jackknife_proportion = 0.8,
  num_N_hypothesized = 11,
  N_try_min = NULL,
  N_try_max = NULL,
  num_p_hypothesized = 20,
  p_hypothesized_range = c(0.01, 0.99),
  n_cores = 1,
  ...
)
}
\description{
Internal handler for the Beta-binomial method with a fixed, global Fis.
}
