% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnadot-beta.R
\name{.dnadot_bhd_internal}
\alias{.dnadot_bhd_internal}
\title{Internal handler for the Buffered Histogram Discrepancy (BHD) method.}
\usage{
.dnadot_bhd_internal(
  genind_object,
  jackknife_proportion = 0.8,
  bhd_buffer_width = 1,
  num_N_hypothesized = 11,
  N_try_min = NULL,
  N_try_max = NULL,
  num_p_hypothesized = 20,
  p_hypothesized_range = c(0.01, 0.99),
  n_cores = 1,
  ...
)
}
\description{
Internal handler for the Buffered Histogram Discrepancy (BHD) method.
}
