% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnadot-beta.R
\name{.dnadot_mde_internal}
\alias{.dnadot_mde_internal}
\title{Internal handler for standard MDE methods (Wasserstein, Sherwin, NLL).}
\usage{
.dnadot_mde_internal(
  genind_object,
  discrepancy_type,
  jackknife_proportion = 0.8,
  num_N_hypothesized = 11,
  N_try_min = NULL,
  N_try_max = NULL,
  num_p_hypothesized = 20,
  p_hypothesized_range = c(0.01, 0.99),
  reg_strength = 0,
  reg_N_threshold = Inf,
  n_cores = 1,
  ...
)
}
\description{
Internal handler for standard MDE methods (Wasserstein, Sherwin, NLL).
}
