% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnadot-beta.R
\name{.dnadot_mle_internal}
\alias{.dnadot_mle_internal}
\title{Internal handler for the Maximum Likelihood Estimation (MLE) method.
NOTE: This method is flagged as experimental and potentially flawed.}
\usage{
.dnadot_mle_internal(
  genind_object,
  jackknife_proportion = 0.8,
  mle_target_alleles = NULL,
  num_N_for_aic = 30,
  N_hypothesized_range_factor = 3,
  n_cores = 1,
  ...
)
}
\description{
Internal handler for the Maximum Likelihood Estimation (MLE) method.
NOTE: This method is flagged as experimental and potentially flawed.
}
