% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnadot-beta.R
\name{.process_results}
\alias{.process_results}
\title{Process and summarize results from all loci.}
\usage{
.process_results(all_loci_results, method_name)
}
\arguments{
\item{all_loci_results}{A list containing the results from each locus-allele combination.}

\item{method_name}{The name of the method used, for reporting.}
}
\value{
A list containing the final Nc estimate and detailed data frames.
}
\description{
This function takes the raw list of results from the `foreach` loop,
filters out errors, selects the most informative loci/alleles, and
calculates the final, single estimate of Nc.
}
