% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnadot-beta.R
\name{.setup_grid_params}
\alias{.setup_grid_params}
\title{Set up the grid search parameters.}
\usage{
.setup_grid_params(
  n_individuals,
  jackknife_proportion,
  num_N_hypothesized,
  N_try_min,
  N_try_max,
  num_p_hypothesized,
  p_hypothesized_range
)
}
\arguments{
\item{n_individuals}{Total number of individuals in the sample.}

\item{jackknife_proportion}{The proportion of individuals to use in each subsample.}

\item{num_N_hypothesized}{The number of Nc values to test.}

\item{N_try_min}{The minimum Nc value to test.}

\item{N_try_max}{The maximum Nc value to test.}

\item{num_p_hypothesized}{The number of p values to test.}

\item{p_hypothesized_range}{A vector with the min and max p to test.}
}
\value{
A list containing the vectors of Nc and p values to try, and the subsample size.
}
\description{
This function defines the grid of hypothesized Nc and allele frequency (p)
values that will be searched to find the best estimate. It also calculates
the size of the jackknife subsamples.
}
