% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnadot-beta.R
\name{.setup_parallel}
\alias{.setup_parallel}
\title{Set up the parallel backend for analysis.}
\usage{
.setup_parallel(n_cores, n_items)
}
\arguments{
\item{n_cores}{Number of cores to use.}

\item{n_items}{The number of items to iterate over (e.g., number of loci).}
}
\value{
A boolean indicating whether the analysis will run in parallel.
}
\description{
This function checks if parallel processing is requested and possible,
and registers the appropriate backend with the `foreach` package.
}
