% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dotR_dnadot_betabin_cpp}
\alias{dotR_dnadot_betabin_cpp}
\title{MDE using Beta-binomial distribution (Fixed Fis)}
\usage{
dotR_dnadot_betabin_cpp(
  genotypes_locus,
  n_individuals,
  n_subsample_individuals,
  N_try_values,
  p_try_values,
  target_allele_idx,
  global_fis
)
}
\description{
MDE using Beta-binomial distribution (Fixed Fis)
}
\keyword{internal}
