% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dotR_dnadot_betabin_dynamic_cpp}
\alias{dotR_dnadot_betabin_dynamic_cpp}
\title{--- NEW FUNCTION ---
MDE using Beta-binomial distribution (Dynamic Fis)}
\usage{
dotR_dnadot_betabin_dynamic_cpp(
  genotypes_locus,
  n_individuals,
  n_subsample_individuals,
  N_try_values,
  p_try_values,
  target_allele_idx,
  migration_rate_m
)
}
\description{
--- NEW FUNCTION ---
MDE using Beta-binomial distribution (Dynamic Fis)
}
\keyword{internal}
