% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dotR_dnadot_bhd_cpp}
\alias{dotR_dnadot_bhd_cpp}
\title{Buffered Histogram Discrepancy (BHD) Method (C++ Core Logic)}
\usage{
dotR_dnadot_bhd_cpp(
  genotypes_locus,
  n_individuals,
  n_subsample_individuals,
  N_try_values,
  p_try_values,
  target_allele_idx,
  bhd_buffer_width
)
}
\description{
Buffered Histogram Discrepancy (BHD) Method (C++ Core Logic)
}
\keyword{internal}
