% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dotR_dnadot_mde_cpp}
\alias{dotR_dnadot_mde_cpp}
\title{Minimum Distance Estimation (MDE) Core Logic (C++ Core Logic)}
\usage{
dotR_dnadot_mde_cpp(
  genotypes_locus,
  n_individuals,
  n_subsample_individuals,
  N_try_values,
  p_try_values,
  target_allele_idx,
  discrepancy_type,
  reg_strength,
  reg_N_threshold,
  debug = FALSE
)
}
\description{
Minimum Distance Estimation (MDE) Core Logic (C++ Core Logic)
}
\keyword{internal}
