% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dotR_hypergeo_cpp}
\alias{dotR_hypergeo_cpp}
\title{Calculate Hypergeometric Probability (Stable Version)}
\usage{
dotR_hypergeo_cpp(k_drawn, n_draws, K_successes_in_pop, N_total_items_in_pop)
}
\description{
Calculate Hypergeometric Probability (Stable Version)
}
\keyword{internal}
