% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnadot-beta.R
\name{read_vcf_to_genind_dotR}
\alias{read_vcf_to_genind_dotR}
\title{Read a VCF file and convert it to a dotR genind object}
\usage{
read_vcf_to_genind_dotR(vcf_file, ploidy = 2)
}
\arguments{
\item{vcf_file}{Path to the VCF file.}

\item{ploidy}{The ploidy of the organism (default is 2).}
}
\value{
A genind object ready for use with `dnadot_snp`.
}
\description{
A convenience wrapper that reads a VCF file, extracts genotype information,
and uses `df2genind_dotR` to create the custom genind object required by the package.
}
